Controllers: double omgevormd naar decimal

+++ src/KairosWeb_Groep6/Controllers/Baten/ExterneInkopenController.cs
@@ -170,7 +170,7 @@ namespace KairosWeb_Groep6.Controllers.Baten
                 ViewData["totaal"] = 0;
             }
 
-            double totaal = analyse.ExterneInkopen
+            decimal totaal = analyse.ExterneInkopen
                                     .Sum(t => t.Bedrag);
 
             ViewData["totaal"] = totaal.ToString("C", new CultureInfo("nl-BE"));
