+++ src/KairosWeb_Groep6/Controllers/Baten/MedewerkersHogerNiveauController.cs
@@ -173,8 +173,8 @@ namespace KairosWeb_Groep6.Controllers.Baten
                 ViewData["totaal"] = 0;
             }
 
-            if(analyse.Departement != null) { 
-                double totaal = MedewerkerNiveauBaatExtensions.GeefTotaal(
+            if(analyse.Departement != null) {
+                decimal totaal = MedewerkerNiveauBaatExtensions.GeefTotaal(
                     analyse.MedewerkersHogerNiveauBaat,
                     analyse.Departement.Werkgever.AantalWerkuren,
                     analyse.Departement.Werkgever.PatronaleBijdrage);
