+++ src/KairosWeb_Groep6/Controllers/Kosten/BegeleidingsKostenController.cs
@@ -182,7 +182,7 @@ namespace KairosWeb_Groep6.Controllers.Kosten
 
             if (analyse.Departement != null)
             {
-                double totaal = BegeleidingsKostExtensions.GeefTotaal(analyse.BegeleidingsKosten,
+                decimal totaal = BegeleidingsKostExtensions.GeefTotaal(analyse.BegeleidingsKosten,
                     analyse.Departement.Werkgever.PatronaleBijdrage);
 
                 ViewData["totaal"] = totaal.ToString("C", new CultureInfo("nl-BE"));
