+++ src/KairosWeb_Groep6/Controllers/Kosten/InfrastructuurKostenController.cs
@@ -174,7 +174,7 @@ namespace KairosWeb_Groep6.Controllers.Kosten
                 ViewData["totaal"] = 0;
             }
 
-            double totaal = KostOfBaatExtensions.GeefTotaal(analyse.InfrastructuurKosten);
+            decimal totaal = KostOfBaatExtensions.GeefTotaal(analyse.InfrastructuurKosten);
 
             ViewData["totaal"] = totaal.ToString("C", new CultureInfo("nl-BE"));
         }
